backend/requirements.txt
django==5.1.6
djangorestframework==3.15.2
psycopg2==2.9.10 ; platform_machine != "aarch64"
psycopg2-binary==2.9.10 ; platform_machine == "aarch64"
whitenoise==6.9.0
web3==7.8.0
